killed tons of warnings made the upgrades last longer if more than one level has to be upgraded

+++ src/GameHints.cpp
@@ -54,7 +54,7 @@ void GameHints::removeHint(int n)
 
 void GameHints::setGameHintText(int n, const std::string& hint)
 {
-	assert (n < texts.size());
+	assert (n < (int)texts.size());
 	texts[n]=hint;
 }
 
@@ -62,7 +62,7 @@ void GameHints::setGameHintText(int n, const std::string& hint)
 
 const std::string& GameHints::getGameHintText(int n)
 {
-	assert (n < texts.size());
+	assert (n < (int)texts.size());
 	return texts[n];
 }
 
@@ -70,7 +70,7 @@ const std::string& GameHints::getGameHintText(int n)
 
 void GameHints::setHintHidden(int n)
 {
-	if (n >= 0 && n < hidden.size())
+	if (n >= 0 && n < (int)hidden.size())
 		hidden[n]=true;
 }
 
@@ -78,7 +78,7 @@ void GameHints::setHintHidden(int n)
 
 void GameHints::setHintVisible(int n)
 {
-	if (n >= 0 && n < hidden.size())
+	if (n >= 0 && n < (int)hidden.size())
 		hidden[n]=false;
 }
 
@@ -86,7 +86,7 @@ void GameHints::setHintVisible(int n)
 
 bool GameHints::isHintVisible(int n)
 {
-	if (n >= 0 && n < hidden.size())
+	if (n >= 0 && n < (int)hidden.size())
 		return !hidden[n];
 	else
 		return false;
@@ -96,7 +96,7 @@ bool GameHints::isHintVisible(int n)
 
 void GameHints::setScriptNumber(int n, int scriptNumber)
 {
-	assert(n < scriptNumbers.size());
+	assert(n < (int)scriptNumbers.size());
 	scriptNumbers[n]=scriptNumber;
 }
 
@@ -104,7 +104,7 @@ void GameHints::setScriptNumber(int n, int scriptNumber)
 
 int GameHints::getScriptNumber(int n)
 {
-	assert(n < scriptNumbers.size());
+	assert(n < (int)scriptNumbers.size());
 	return scriptNumbers[n];
 }
 
@@ -114,7 +114,7 @@ void GameHints::encodeData(GAGCore::OutputStream* stream) const
 {
 	stream->writeEnterSection("GameHints");
 	stream->writeUint32(texts.size(), "size");
-	for(int i=0; i<texts.size(); ++i)
+	for(unsigned int i=0; i<texts.size(); ++i)
 	{
 		stream->writeEnterSection(i);
 		stream->writeText(texts[i], "text");
@@ -134,7 +134,7 @@ void GameHints::decodeData(GAGCore::InputStream* stream, Uint32 versionMinor)
 	scriptNumbers.clear();
 	stream->readEnterSection("GameHints");
 	Uint32 size = stream->readUint32("size");
-	for(int i=0; i<size; ++i)
+	for(unsigned int i=0; i<size; ++i)
 	{
 		stream->readEnterSection(i);
 		texts.push_back(stream->readText("text"));
